fix: update directory and test file name
