+++ src/io/tradetrust/invoice/1.0/Invoice.v3.json
