+++ src/io/tradetrust/invoice/1.0/context.test.ts
@@ -1,5 +1,3 @@
-import CertificateOfOriginSchema from "./CertificateOfOrigin.v3.json";
-import CertificateOfOriginSampleData from "./certificate-of-origin-data.json";
 import InvoiceSchema from "./Invoice.v3.json";
 import InvoiceSampleData from "./invoice-data.json";
 import { expand, JsonLdDocument } from "jsonld";
@@ -18,34 +16,6 @@ const expandDocument = async (mergedDocument: JsonLdDocument) => {
     }
   });
 };
-
-describe("certificate of origin schema", () => {
-  it("should work with valid certificate of origin", async () => {
-    const mergedDocument = {
-      ...CertificateOfOriginSchema,
-      ...CertificateOfOriginSampleData
-    } as JsonLdDocument;
-    expect(await expandDocument(mergedDocument)).toBeTruthy();
-  });
-
-  it("should throw error when unknown context or properties found", async () => {
-    const modifiedData = {
-      ...CertificateOfOriginSampleData,
-      invalidCOOProperties: "Random String"
-    };
-
-    const mergedDocument = {
-      ...CertificateOfOriginSchema,
-      ...modifiedData
-    } as JsonLdDocument;
-
-    return expandDocument(mergedDocument).catch(e => {
-      expect(e.message).toMatchInlineSnapshot(
-        `"\\"The property invalidCOOProperties in the input was not defined in the context\\""`
-      );
-    });
-  });
-});
 
 describe("invoice schema", () => {
   it("should work with valid invoice", async () => {
